rls is now enforced on expenses table
